requirements.txt
# Twitter API
tweepy==4.14.0

# Database
SQLAlchemy==2.0.23

# Environment variables
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0

# RSS Feed Parser
feedparser==6.0.10

# Data processing
pandas==2.1.4

# HTTP requests
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==6.1.0

# Logging
colorlog==6.8.0

# Date/time handling
python-dateutil==2.8.2
